@@ -1,42 +1,40 @@
 {
-  "name": "Djini",
+  "name": "djini",
   "version": "1.0.6",
   "private": true,
   "scripts": {
+    "postinstall": "rm ./node_modules/react-native/local-cli/core/__fixtures__/files/package.json",
     "start": "node node_modules/react-native/local-cli/cli.js start",
     "test": "jest --no-cache",
     "test:watch": "npm run test -- --watch"
-  },
-  "jest": {
-    "preset": "jest-react-native"
   },
   "dependencies": {
     "google-libphonenumber": "^2.0.3",
     "immutable": "^3.8.1",
     "moment": "^2.15.0",
-    "parse": "^1.9.2",
-    "react": "15.3.1",
-    "react-native": "^0.34.0",
-    "react-native-code-push": "^1.14.6-beta",
-    "react-native-contacts": "^0.5.1",
-    "react-native-image-picker": "^0.22.7",
-    "react-native-linear-gradient": "^1.5.12",
-    "react-native-push-notification": "^2.1.1",
-    "react-native-router-flux": "3.35.0",
-    "react-native-swipeout": "dancormier/react-native-swipeout",
-    "react-native-vector-icons": "^2.1.0",
-    "react-redux": "^4.4.5",
-    "redux": "^3.5.2",
-    "redux-logger": "^2.6.1",
-    "redux-thunk": "^2.1.0"
+    "parse": "^1.11.0",
+    "react": "^16.2.0",
+    "react-native": "0.52.2",
+    "react-native-code-push": "5.2.1",
+    "react-native-contacts": "2.1.1",
+    "react-native-image-picker": "^0.26.7",
+    "react-native-linear-gradient": "^2.4.0",
+    "react-native-push-notification": "^3.0.2",
+    "react-native-router-flux": "3.43.0",
+    "react-native-swipeout": "2.3.3",
+    "react-native-vector-icons": "^4.5.0",
+    "react-redux": "^5.0.6",
+    "redux": "^3.7.2",
+    "redux-logger": "^3.0.6",
+    "redux-thunk": "^2.2.0"
   },
   "devDependencies": {
-    "babel-eslint": "^6.1.2",
-    "babel-jest": "^14.1.0",
-    "babel-preset-react-native": "^1.9.0",
-    "eslint": "^3.4.0",
-    "eslint-plugin-react": "^6.2.0",
-    "jest": "^14.1.0",
-    "jest-react-native": "^14.1.3"
+    "babel-jest": "^22.1.0",
+    "babel-preset-react-native": "4.0.0",
+    "jest": "22.1.4",
+    "react-test-renderer": "16.2.0"
+  },
+  "jest": {
+    "preset": "react-native"
   }
 }
