@@ -8,10 +8,9 @@
  */
 
 #import "AppDelegate.h"
-#import "CodePush.h"
-
-#import "RCTBundleURLProvider.h"
-#import "RCTRootView.h"
+#import <CodePush/CodePush.h>
+#import <React/RCTBundleURLProvider.h>
+#import <React/RCTRootView.h>
 #import "RCTPushNotificationManager.h"
 
 @implementation AppDelegate
