@@ -1,4 +1,5 @@
-import React, {Component, PropTypes} from 'react'
+import PropTypes from 'prop-types';
+import React, { Component } from 'react';
 import {View, StyleSheet} from 'react-native'
 import moment from 'moment'
 import DjiniTextInput from './DjiniTextInput'
@@ -148,5 +149,3 @@ const styles = StyleSheet.create({
     marginRight: 0,
   }
 })
-
-
