@@ -1,4 +1,5 @@
-import React, {PropTypes} from 'react'
+import PropTypes from 'prop-types';
+import React from 'react';
 import {View, StyleSheet} from 'react-native'
 import DjiniText from './DjiniText'
 import DjiniButton from './DjiniButton'
