@@ -1,4 +1,5 @@
-import React, {PropTypes, Component} from 'react'
+import PropTypes from 'prop-types';
+import React, { Component } from 'react';
 import {Text, StyleSheet, Platform} from 'react-native'
 import ViewMoreText from './ViewMoreText'
 
