@@ -1,6 +1,7 @@
 /* global setTimeout */
 import { connect } from 'react-redux'
-import React, {Component, PropTypes} from 'react'
+import React, {Component} from 'react'
+import PropTypes from 'prop-types'
 import {StyleSheet} from 'react-native'
 
 import DjiniText from './DjiniText'
