@@ -1,4 +1,5 @@
-import React, {PropTypes} from 'react'
+import PropTypes from 'prop-types';
+import React from 'react';
 import {StyleSheet, View, InteractionManager} from 'react-native'
 import {Actions} from 'react-native-router-flux'
 
