@@ -1,5 +1,6 @@
 import Immutable from 'immutable';
-import React, {Component, PropTypes} from 'react';
+import PropTypes from 'prop-types';
+import React, { Component } from 'react';
 import {StyleSheet, View, ListView, TouchableHighlight, TouchableOpacity, Text} from 'react-native';
 
 // actions
