@@ -1,6 +1,6 @@
-
 import { connect } from 'react-redux';
-import React, {Component, PropTypes} from 'react'
+import PropTypes from 'prop-types';
+import React, { Component } from 'react';
 import { StyleSheet, View} from 'react-native';
 import {Actions} from 'react-native-router-flux'
 
