@@ -13,7 +13,7 @@
  */
 import { createStore, applyMiddleware } from 'redux'
 import thunk from 'redux-thunk'
-import createLogger from 'redux-logger'
+import {createLogger} from 'redux-logger'
 import {enableReduxLogger} from './config'
 
 /**
